samurai-examples/testcase/testcase/www/html/testcases/css3/device-adapt/viewport-initial-height-zero-expected.txt

ALERT: viewport size 200x100 scale 1 with limits [1, 5] and userScalable true
